[wip][openmp] Fix taskgraph tdg_id to not function as an idx
Tdg_id was implementatd as an idx, this implementation needs an
update in order to match the OpenMP spcs 6.0, were a graph_id
clause have been introduced. This MR assumes that the tdg_id is
generatd in LLVM CG and the runtime just gets a hashed id were
region/graph_id have taken into account.

This chenge is backwards compatible