[OpenMP] Fix tdg_id in task record and replay mechanism
Previously, tdg_id was used as an array index and an id for the task replay
and record mechanism, but to adder to the OpenMP 6.0 specifications it
should be trated uniquly as a id. This makes the mechanism fully
compliant with OpenMP 6.0 graph_id clause without breaking
compatibility.

With this change, tdg_id more accurately represents an opaque
identifier, accommodating the scenario where the compiler generates a
hashed ID that accounts for the region/graph ID.